Relocating bit converter to shared code.
