+++ shared/BitConverter.cpp
