+++ shared/BitConverter.h
